Add testable example for record user signup/singin using custom struct
Ref #235
